refactor(ui): drop the liquid glass expect/actual so Desktop renders it too
