--- composeApp/src/androidMain/kotlin/com/maxrave/simpmusic/expect/ui/LiquidGlass.android.kt
@@ -1,68 +1,0 @@
-package com.maxrave.simpmusic.expect.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.layer.GraphicsLayer
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.util.lerp
-import com.kyant.backdrop.backdrops.LayerBackdrop
-import com.kyant.backdrop.backdrops.rememberLayerBackdrop
-import com.kyant.backdrop.drawBackdrop
-import com.kyant.backdrop.effects.blur
-import com.kyant.backdrop.effects.colorControls
-import com.kyant.backdrop.effects.lens
-import com.kyant.backdrop.effects.vibrancy
-import com.maxrave.simpmusic.ui.theme.LocalIsDarkTheme
-import kotlin.math.sign
-import com.kyant.backdrop.backdrops.layerBackdrop as nativeBackdrop
-
-actual typealias PlatformBackdrop = LayerBackdrop
-
-@Composable
-actual fun rememberBackdrop(color: Color): PlatformBackdrop =
-    rememberLayerBackdrop {
-        drawRect(color)
-        drawContent()
-    }
-
-actual fun Modifier.layerBackdrop(backdrop: PlatformBackdrop): Modifier = this.nativeBackdrop(backdrop)
-
-actual fun Modifier.drawBackdropCustomShape(
-    backdrop: PlatformBackdrop,
-    layer: GraphicsLayer,
-    luminanceAnimation: Float,
-    shape: Shape,
-): Modifier =
-    this.drawBackdrop(
-        backdrop = backdrop,
-        effects = {
-            val l = (luminanceAnimation * 2f - 1f).let { sign(it) * it * it }
-            vibrancy()
-            colorControls(
-                // Neutral brightness/contrast: the old curve washed the glass out to white on bright
-                // backgrounds ("đục trắng"). Darkening is done in onDrawSurface (keeps "đục đen").
-                brightness = 0.05f,
-                contrast = 1f,
-                saturation = 1.5f,
-            )
-            blur(
-                if (l > 0f) {
-                    lerp(8f.dp.toPx(), 16f.dp.toPx(), l)
-                } else {
-                    lerp(8f.dp.toPx(), 2f.dp.toPx(), -l)
-                },
-            )
-            lens(24f.dp.toPx(), size.minDimension / 2f, true)
-        },
-        onDrawBackdrop = { drawBackdrop ->
-            drawBackdrop()
-            layer.record { drawBackdrop() }
-        },
-        shape = { shape },
-        onDrawSurface = {
-            val darken = lerp(0.12f, 0.5f, ((luminanceAnimation - 0.3f) / 0.5f).coerceIn(0f, 1f))
-            drawRect(Color.Black.copy(alpha = darken))
-        },
-    )
