+++ composeApp/src/commonMain/kotlin/com/maxrave/simpmusic/ui/component/LiquidGlassContainer.kt
@@ -1,68 +1,126 @@
 package com.maxrave.simpmusic.ui.component
 
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.spring
+import androidx.compose.foundation.gestures.awaitEachGesture
+import androidx.compose.foundation.gestures.awaitFirstDown
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.BlendMode
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.graphics.layer.GraphicsLayer
+import androidx.compose.ui.graphics.rememberGraphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.input.pointer.AwaitPointerEventScope
+import androidx.compose.ui.input.pointer.PointerEventPass
+import androidx.compose.ui.input.pointer.PointerId
+import androidx.compose.ui.input.pointer.PointerInputChange
+import androidx.compose.ui.input.pointer.PointerInputScope
+import androidx.compose.ui.input.pointer.changedToUpIgnoreConsumed
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.input.pointer.positionChange
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.fastFirstOrNull
+import androidx.compose.ui.util.lerp
+import com.kyant.backdrop.drawBackdrop
+import com.kyant.backdrop.effects.blur
+import com.kyant.backdrop.effects.colorControls
+import com.kyant.backdrop.effects.lens
+import com.kyant.backdrop.effects.vibrancy
+import com.kyant.backdrop.highlight.Highlight
 import com.maxrave.simpmusic.expect.ui.PlatformBackdrop
 import com.maxrave.simpmusic.ui.theme.LocalIsDarkTheme
-import org.jetbrains.compose.resources.DrawableResource
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+import kotlin.math.sign
 
 /**
  * Applies the SimpMusic liquid-glass effect to any element.
  *
- * This is the single platform-specific primitive behind the glass buttons that
- * used to be hand-wired inline on every screen. It encapsulates the per-surface
+ * This is the single primitive behind the glass buttons that used to be hand-wired
+ * inline on every screen. It encapsulates the per-surface
  * [androidx.compose.ui.graphics.layer.GraphicsLayer], the Kyant `drawBackdrop`
- * effect stack and — on Android — the press/hold "liquid" interaction (a slight
- * scale-up, deeper refraction and a radial glow that follows the finger, springing
- * back on release). The press gesture is observe-only, so wrapped click handlers
- * keep working.
+ * effect stack and the press/hold "liquid" interaction (a slight scale-up, deeper
+ * refraction and a radial glow that follows the pointer, springing back on release).
+ * The press gesture is observe-only, so wrapped click handlers keep working.
  *
  * The element MUST be a sibling of the backdrop source (the box carrying
  * [com.maxrave.simpmusic.expect.ui.layerBackdrop]); nesting it inside the source
  * creates a render-feedback loop that crashes the RuntimeShader.
  *
- * On desktop (JVM) the underlying backdrop modifiers are no-ops, so this only
- * clips to [shape] and the look degrades gracefully.
+ * Runs on Android and desktop alike: Kyant's backdrop is a KMP artifact whose desktop
+ * variant renders through skiko. The press interaction is driven by `pointerInput`, so
+ * a mouse button triggers it exactly like a finger.
  *
- * @param interactive Android only — set false for a static glass surface.
+ * @param interactive set false for a static glass surface.
  */
 @Composable
-expect fun Modifier.liquidGlass(
+fun Modifier.liquidGlass(
     backdrop: PlatformBackdrop,
     shape: Shape = CircleShape,
     interactive: Boolean = true,
-): Modifier
+    highlight: Highlight = Highlight.Default,
+): Modifier {
+    val isDark = LocalIsDarkTheme.current
+    val layer = rememberGraphicsLayer()
+    val interaction = rememberGlassInteraction()
+    return this.drawInteractiveGlass(
+        isDark = isDark,
+        backdrop = backdrop,
+        layer = layer,
+        luminanceAnimation = 0.5f,
+        shape = shape,
+        interaction = if (interactive) interaction else null,
+        highlight = highlight,
+    )
+}
 
 /**
  * Overload of [liquidGlass] for surfaces that sample their own background luminance
  * (e.g. the MiniPlayer and the bottom bar capsule): the caller owns the [layer] the
  * glass records into and drives [luminanceAnimation], so the glass keeps adapting to
  * the content behind it — unlike the [liquidGlass] above, which uses a fixed
- * mid-luminance. On Android it adds the same press/hold interaction; on desktop it
- * degrades to a plain clip.
+ * mid-luminance.
  */
 @Composable
-expect fun Modifier.liquidGlass(
+fun Modifier.liquidGlass(
     backdrop: PlatformBackdrop,
     layer: GraphicsLayer,
     luminanceAnimation: Float,
     shape: Shape = CircleShape,
     interactive: Boolean = true,
-): Modifier
+): Modifier {
+    val isDark = LocalIsDarkTheme.current
+    val interaction = rememberGlassInteraction()
+    return this.drawInteractiveGlass(
+        isDark = isDark,
+        backdrop = backdrop,
+        layer = layer,
+        luminanceAnimation = luminanceAnimation,
+        shape = shape,
+        interaction = if (interactive) interaction else null,
+        // MiniPlayer (the only caller of this layer + luminance overload) is a wide surface, so the
+        // shared 1.12 press scale bulges too hard; use a gentler scale here.
+        pressedScale = 1.04f,
+    )
+}
 
 /**
  * A liquid-glass surface wrapping arbitrary [content] (e.g. a pill of icon
- * buttons). Thin convenience over [liquidGlass]; pure common code.
+ * buttons). Thin convenience over [liquidGlass].
  */
 @Composable
 fun LiquidGlassContainer(
@@ -70,11 +128,12 @@ fun LiquidGlassContainer(
     modifier: Modifier = Modifier,
     shape: Shape = CircleShape,
     interactive: Boolean = true,
+    highlight: Highlight = Highlight.Default,
     contentAlignment: Alignment = Alignment.Center,
     content: @Composable BoxScope.() -> Unit,
 ) {
     Box(
-        modifier = modifier.liquidGlass(backdrop, shape, interactive),
+        modifier = modifier.liquidGlass(backdrop, shape, interactive, highlight),
         contentAlignment = contentAlignment,
         content = content,
     )
@@ -92,6 +151,7 @@ fun LiquidGlassIconButton(
     shape: Shape = CircleShape,
     tint: Color = Color.White,
     interactive: Boolean = true,
+    highlight: Highlight = Highlight.Default,
     onClick: () -> Unit,
 ) {
     LiquidGlassContainer(
@@ -99,11 +159,229 @@ fun LiquidGlassIconButton(
         modifier = modifier,
         shape = shape,
         interactive = interactive,
+        highlight = highlight,
     ) {
         RippleIconButton(
             imageVector = imageVector,
             tint = tint,
             onClick = onClick,
         )
+    }
+}
+
+/**
+ * Press/hold state holder for a single liquid-glass surface.
+ *
+ * Project-local, dependency-light reimplementation of Kyant's catalog
+ * `InteractiveHighlight`: instead of the library-internal AGSL highlight shader
+ * (whose public helpers are not exposed in backdrop 2.0.0) we drive a plain
+ * [Brush.radialGradient] from [touchPosition] and a spring-animated
+ * [pressProgress]. The drag detection is observe-only so wrapped buttons keep
+ * receiving their own clicks.
+ */
+class GlassInteraction(
+    private val animationScope: CoroutineScope,
+) {
+    private val pressSpec = spring(dampingRatio = 0.5f, stiffness = 300f, visibilityThreshold = 0.001f)
+    private val pressAnimation = Animatable(0f, 0.001f)
+
+    /** 0f at rest, animating to 1f while pressed. Read in draw/effect/layer blocks. */
+    val pressProgress: Float get() = pressAnimation.value
+
+    /** Local-space touch point used as the centre of the press glow. */
+    var touchPosition by mutableStateOf(Offset.Zero)
+        private set
+
+    suspend fun detectPress(pointer: PointerInputScope) =
+        with(pointer) {
+            inspectDragGestures(
+                onDragStart = { down ->
+                    touchPosition = down.position
+                    animationScope.launch { pressAnimation.animateTo(1f, pressSpec) }
+                },
+                onDragEnd = { animationScope.launch { pressAnimation.animateTo(0f, pressSpec) } },
+                onDragCancel = { animationScope.launch { pressAnimation.animateTo(0f, pressSpec) } },
+            ) { change, _ ->
+                touchPosition = change.position
+            }
+        }
+}
+
+@Composable
+fun rememberGlassInteraction(): GlassInteraction {
+    val scope = rememberCoroutineScope()
+    return remember(scope) { GlassInteraction(scope) }
+}
+
+/**
+ * Draws the liquid-glass effect with the same look as the legacy
+ * `drawBackdropCustomShape`, plus an optional press response driven by
+ * [interaction]: the surface scales up a touch, the refraction/blur deepen and a
+ * radial glow follows the pointer. Pass `interaction = null` for a static surface.
+ *
+ * [luminanceAnimation] keeps the brightness/contrast curve of the original
+ * wrapper (the bottom navigation bar animates it; static surfaces pass `0.5f`).
+ */
+fun Modifier.drawInteractiveGlass(
+    isDark: Boolean,
+    backdrop: PlatformBackdrop,
+    layer: GraphicsLayer,
+    luminanceAnimation: Float,
+    shape: Shape,
+    interaction: GlassInteraction?,
+    pressedScale: Float = 1.12f,
+    highlight: Highlight = Highlight.Default,
+): Modifier =
+    this
+        .drawBackdrop(
+            backdrop = backdrop,
+            shape = { shape },
+            // Kyant's own default is Highlight.Default, whose HighlightStyle.Default is a
+            // DIRECTIONAL rim (angle 45°, falloff 1f) — it lights one side rather than the
+            // whole outline. That reads well on an elongated pill, whose long edge catches
+            // the sweep, and is nearly invisible on a small circle. Pass Highlight.Plain for
+            // a uniform rim all the way round.
+            highlight = { highlight },
+            effects = {
+                val l = (luminanceAnimation * 2f - 1f).let { sign(it) * it * it }
+                val press = interaction?.pressProgress ?: 0f
+                vibrancy()
+                colorControls(
+                    // Neutral brightness/contrast: the old curve brightened + washed the glass out
+                    // to white on bright backgrounds ("đục trắng"). Darkening is done in onDrawSurface.
+                    brightness = 0.05f,
+                    contrast = 1f,
+                    saturation = 1.5f,
+                )
+                blur(
+                    (
+                        if (l > 0f) {
+                            lerp(8f.dp.toPx(), 16f.dp.toPx(), l)
+                        } else {
+                            lerp(8f.dp.toPx(), 2f.dp.toPx(), -l)
+                        }
+                    ) + 2f.dp.toPx() * press,
+                )
+                // refractionHeight stays below the stadium inradius (minDimension / 2) so the
+                // top and bottom refraction never meet at the medial axis — that meeting point on
+                // a wide pill is what produced the dark horizontal seam. depthEffect is off to
+                // match the crisp Kyant demo look and avoid the radial discontinuity at the centre.
+                lens(size.minDimension / 4f + 2f.dp.toPx() * press, size.minDimension / 2f, false)
+            },
+            onDrawBackdrop = { drawBackdrop ->
+                drawBackdrop()
+                layer.record { drawBackdrop() }
+            },
+            onDrawSurface = {
+                // Stay "đục đen": darken more as the background brightens so the glass never washes
+                // out to white (shared by the bottom bar capsule, search FAB and detail-screen pills).
+                val darken = lerp(0.12f, 0.5f, ((luminanceAnimation - 0.3f) / 0.5f).coerceIn(0f, 1f))
+                drawRect((if (isDark) Color.Black else Color.White).copy(alpha = darken))
+                val press = interaction?.pressProgress ?: 0f
+                if (press > 0f) {
+                    drawRect(
+                        brush =
+                            Brush.radialGradient(
+                                colors =
+                                    listOf(
+                                        Color.White.copy(alpha = 0.18f * press),
+                                        Color.Transparent,
+                                    ),
+                                center = interaction?.touchPosition ?: Offset(size.width / 2f, size.height / 2f),
+                                radius = size.minDimension * 1.2f,
+                            ),
+                        blendMode = BlendMode.Plus,
+                    )
+                }
+            },
+            layerBlock =
+                if (interaction != null) {
+                    {
+                        val scale = lerp(1f, pressedScale, interaction.pressProgress)
+                        scaleX = scale
+                        scaleY = scale
+                    }
+                } else {
+                    null
+                },
+        ).then(
+            if (interaction != null) {
+                Modifier.pointerInput(interaction) { interaction.detectPress(this) }
+            } else {
+                Modifier
+            },
+        )
+
+/**
+ * Observe-only drag/press recogniser ported from Kyant's catalog
+ * `DragGestureInspector`. It never consumes events, so a glass surface can react
+ * to a press while the buttons it wraps still handle their own taps.
+ */
+internal suspend fun PointerInputScope.inspectDragGestures(
+    onDragStart: (down: PointerInputChange) -> Unit = {},
+    onDragEnd: (change: PointerInputChange) -> Unit = {},
+    onDragCancel: () -> Unit = {},
+    onDrag: (change: PointerInputChange, dragAmount: Offset) -> Unit,
+) {
+    awaitEachGesture {
+        awaitFirstDown(requireUnconsumed = false, pass = PointerEventPass.Initial)
+
+        val down = awaitFirstDown(requireUnconsumed = false)
+
+        onDragStart(down)
+        onDrag(down, Offset.Zero)
+        val upEvent =
+            drag(
+                pointerId = down.id,
+                onDrag = { onDrag(it, it.positionChange()) },
+            )
+        if (upEvent == null) {
+            onDragCancel()
+        } else {
+            onDragEnd(upEvent)
+        }
+    }
+}
+
+private suspend inline fun AwaitPointerEventScope.drag(
+    pointerId: PointerId,
+    onDrag: (PointerInputChange) -> Unit,
+): PointerInputChange? {
+    val isPointerUp = currentEvent.changes.fastFirstOrNull { it.id == pointerId }?.pressed != true
+    if (isPointerUp) {
+        return null
+    }
+    var pointer = pointerId
+    while (true) {
+        val change = awaitDragOrUp(pointer) ?: return null
+        if (change.isConsumed) {
+            return null
+        }
+        if (change.changedToUpIgnoreConsumed()) {
+            return change
+        }
+        onDrag(change)
+        pointer = change.id
+    }
+}
+
+private suspend inline fun AwaitPointerEventScope.awaitDragOrUp(pointerId: PointerId): PointerInputChange? {
+    var pointer = pointerId
+    while (true) {
+        val event = awaitPointerEvent()
+        val dragEvent = event.changes.fastFirstOrNull { it.id == pointer } ?: return null
+        if (dragEvent.changedToUpIgnoreConsumed()) {
+            val otherDown = event.changes.fastFirstOrNull { it.pressed }
+            if (otherDown == null) {
+                return dragEvent
+            } else {
+                pointer = otherDown.id
+            }
+        } else {
+            val hasDragged = dragEvent.previousPosition != dragEvent.position
+            if (hasDragged) {
+                return dragEvent
+            }
+        }
     }
 }
